requirements.txt
-e .

albumentations
ipywidgets
joblib
jupyterlab
manim
matplotlib
natsort
netron
numpy
opencv-contrib-python
optuna
pandas
pillow
lightning[extra]
raster_geometry
SciencePlots
scikit-learn
scipy
screeninfo
seaborn
tables
# torchinfo
tensorboard
tensorflow
torchmetrics[detection]
tqdm
typing-extensions

--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
